Fix failure to apply filter when searching in chosen
A classic case of ordering differing between browsers, apparently. Affected
Firefox and Safari but not Chrome.

Original problem description:

  start typing a muscle name  e.g., medial pterygoid muscle,  the text will
  auto-complete and then select this name by left clicking on the name

  NOTHING WILL HAPPEN and the box will go blank

In FF and Safari, my guess is that the form was being submitted by this handler
before the form contents could be updated. Perhaps the text area used for
typing was considered "changed" (thus triggering this handler) before Chosen
had a chance to update the select box.